gender-package {gender}	R Documentation

Gender: predict gender by name from historical data

Description

Gender: predict gender from names using historical data

Details

Encodes gender based on names and dates of birth, using U.S. Census or Social Security data sets. Requires separate downlaod of datasets, which should be done automatically and can be done manually by running install_genderdata_package().

Author(s)

lincoln@lincolnmullen.com

[Package gender version 0.5.2 Index]