install_genderdata_package {gender}	R Documentation

Install the genderdata package after checking with the user

Description

Install the genderdata package after checking with the user

Usage

install_genderdata_package()

[Package gender version 0.5.2 Index]