check_genderdata_package {gender}	R Documentation

Check whether to install data for gender function and install if necessary

Description

If the genderdata package is not installed, install it from GitHub using devtools. If it is not up to date, reinstall it.

Usage

check_genderdata_package()

[Package gender version 0.5.2 Index]